commit all

+++ src/main/java/peaksoft/repositories/repoImpl/InstructorRepoImpl.java
@@ -1,6 +1,8 @@
 package peaksoft.repositories.repoImpl;
 
 import jakarta.persistence.EntityManager;
+import jakarta.persistence.EntityTransaction;
+import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.hibernate.HibernateException;
 import org.springframework.stereotype.Repository;
@@ -8,13 +10,14 @@ import peaksoft.entities.Course;
 import peaksoft.entities.Instructor;
 import peaksoft.repositories.InstructorRepo;
 
+import java.util.ArrayList;
 import java.util.List;
 
 @Repository
 @RequiredArgsConstructor
 public class InstructorRepoImpl implements InstructorRepo {
     private final EntityManager entityManager;
-
+@Transactional
     @Override
     public Instructor getById(Long id) {
         try {
@@ -30,28 +33,26 @@ public class InstructorRepoImpl implements InstructorRepo {
 
     @Override
     public List<Instructor> getAll() {
-        try {
             entityManager.getTransaction().begin();
-            List<Instructor> instructors = entityManager.createQuery("select i from Instructor i").getResultList();
+            List<Instructor> instructors = entityManager.createQuery("select i from Instructor i",Instructor.class).getResultList();
             entityManager.getTransaction().commit();
             return instructors;
-        }catch (HibernateException e){
-            System.out.println(e.getMessage());
         }
-        return null;
-    }
+
 
     @Override
-    public void save(Instructor instructor) {
+    public void save(Instructor instructor, Long courseId) {
         try {
             entityManager.getTransaction().begin();
+            Course course = entityManager.find(Course.class, courseId);
             entityManager.persist(instructor);
+            course.getInstructors().add(instructor);
+            instructor.getCourses().add(course);
             entityManager.getTransaction().commit();
         }catch (HibernateException e){
             System.out.println(e.getMessage());
         }
     }
-
     @Override
     public void update(Long id, Instructor newInstructor) {
         try {
@@ -72,27 +73,59 @@ public class InstructorRepoImpl implements InstructorRepo {
 
     @Override
     public void delete(Long id) {
-        try {
-            entityManager.getTransaction().begin();
+    entityManager.getTransaction().begin();
+
             Instructor instructor = entityManager.find(Instructor.class, id);
-            entityManager.remove(instructor);
+
+            if (instructor != null) {
+                for (Course course : instructor.getCourses()) {
+                    if (course != null) {
+                        course.getInstructors().remove(instructor);
+                    }
+                }
+                entityManager.remove(instructor);
+            }
             entityManager.getTransaction().commit();
             System.out.println("Instructor with id " + id + " deleted successfully");
-        }catch (HibernateException e){
-            System.out.println(e.getMessage());
+
         }
-    }
+
+
     @Override
-    public List<Instructor>getInstructorsByCourse(Long courseId) {
+    public void assignInstructorToCourse(Long courseId, Long instructorId) {
         try {
             entityManager.getTransaction().begin();
             Course course = entityManager.find(Course.class, courseId);
-            course.getInstructors().add(new Instructor());
+            Instructor instructor = entityManager.find(Instructor.class, instructorId);
+            instructor.setCourses(List.of(course));
+            course.setInstructors(List.of(instructor));
+            entityManager.merge(course);
+            entityManager.merge(instructor);
             entityManager.getTransaction().commit();
-            return course.getInstructors();
         }catch (HibernateException e){
             System.out.println(e.getMessage());
         }
-        return null;
     }
+
+    @Override
+    public List<Instructor> getAllInstructorById(Long courseId) {
+        entityManager.getTransaction().begin();
+        Course course = entityManager.find(Course.class, courseId);
+        List<Instructor> instructors = course.getInstructors();
+        entityManager.getTransaction().commit();
+        return instructors;
+    }
+
+//    @Override
+//    public List<Instructor> getAllInstructorById(Long courseId) {
+//        entityManager.getTransaction().begin();
+//        List<Instructor> instructors = new ArrayList<>();
+//        Course course = entityManager.find(Course.class, courseId);
+//        if (course.getInstructors() != null) {
+//            instructors.addAll(course.getInstructors());
+//        }
+//        entityManager.getTransaction().commit();
+//        return instructors;
+//    }
+
 }
